requirements.txt
streamlit
pandas
numpy
scikit-learn
joblib
nltk
matplotlib
newspaper3k
lxml